training works now

+++ preprocessing/generate_empty_emb.py
@@ -5,6 +5,6 @@ assert pths, "No latent pth found"
 sample = torch.load(pths[0], weights_only=False)
 emb = sample["text_emb"]
 assert emb is not None, "Sample text_emb is None; ensure metadata tasks/text exists during extraction"
-empty = torch.zeros_like(emb)
+empty = torch.zeros_like(emb,dtype=torch.bfloat16)
 torch.save(empty, os.path.join(dataset_root, "empty_emb.pt"))
 print("saved:", os.path.join(dataset_root, "empty_emb.pt"), "shape:", tuple(empty.shape), "dtype:", empty.dtype)
